requirements.txt
altair==5.2.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
build==0.9.0
cachetools==5.3.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
comm==0.2.1
cryptography==3.4.6
debugpy==1.8.0
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multipart==0.2.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
openpyxl==3.1.2
packaging==21.3
pandas==2.2.0
parso==0.8.3
pdfservices-sdk==2.3.0
pep517==0.13.0
pillow==10.2.0
platformdirs==4.2.0
polling==0.3.2
polling2==0.5.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.14.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
pywin32==306
PyYAML==6.0
pyzmq==25.1.2
referencing==0.32.1
regex==2023.12.25
requests==2.27.1
requests-toolbelt==0.10.1
rich==13.7.0
rpds-py==0.17.1
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.30.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.15.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==1.26.13
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.13
XlsxWriter==3.1.9
zipp==3.17.0
